moved copying of source files to binary, + created a packedbinary to provide the old functionality of binary (used by ORCA)
