+++ easybuild/easyconfigs/o/ORCA/ORCA-2_9_1-linux_x86-64.eb
@@ -1,4 +1,4 @@
-easyblock="Binary"
+easyblock="PackedBinary"
 
 name="ORCA"
 version='2_9_1-linux_x86-64'
